Add texture size to voxel format, allows mixing hi/lo-res textures
Now the texture size is encoded in the voxel format (well, the logarithm
of the texture size, since it is always a power of two, so it can occupy
fewer bits - 2^(2^3) for up to 256, or 2^(2^4) for up to 65k), which
ends up in an attribute, instead of globally set in a uniform - so
different textures can have different sizes.

Requires corresponding voxel-shader change.